requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pytest==7.4.3
httpx==0.25.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pytest-asyncio==0.21.1
bcrypt==4.0.1
